refactor: RouteClient 인터페이스화
