+++ backend/src/main/java/com/ody/route/config/RouteConfig.java
@@ -1,5 +1,6 @@
 package com.ody.route.config;
 
+import com.ody.route.service.OdsayRouteClient;
 import com.ody.route.service.RouteClient;
 import java.time.Duration;
 import lombok.RequiredArgsConstructor;
@@ -9,6 +10,7 @@ import org.springframework.boot.web.client.ClientHttpRequestFactorySettings;
 import org.springframework.boot.web.client.RestClientCustomizer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 import org.springframework.http.client.ClientHttpRequestFactory;
 import org.springframework.web.client.RestClient;
 
@@ -20,9 +22,10 @@ public class RouteConfig {
     private final RouteProperties routeProperties;
 
     @Bean
+    @Profile("!test")
     public RouteClient routeClient(RestClient.Builder routeRestClientBuilder) {
 //        routeRestClientCustomizer().customize(routeRestClientBuilder);
-        return new RouteClient(routeProperties, routeRestClientBuilder);
+        return new OdsayRouteClient(routeProperties, routeRestClientBuilder);
     }
 
     @Bean
